requirements.txt
openai-whisper
deep-translator
argostranslate
ffmpeg-python
